refactor(via): remove QueryParamIter in favor of into_vec
